update: successful local test of knowledge base routes, pipeline routes incorporated into server.py but need debugging
